refactor: minor refactor
